new changes

+++ src/controllers/groupmessaging/GroupActions.java
@@ -1,5 +1,10 @@
 package controllers.groupmessaging;
 
+/**
+ * @author : DABAGIRE Valens
+ * @description : Handling group messaging actions
+ */
+
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dbconnection.OnlineDbConnection;
@@ -7,7 +12,6 @@ import models.ResponseStatus;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.util.Iterator;
 import java.util.Map;
 
@@ -87,6 +91,39 @@ public class GroupActions {
         return new ObjectMapper().writeValueAsString(responseStatus);
     }
 
+    public String leaveGroup(JsonNode requestBody) throws Exception{
+
+        ResponseStatus responseStatus = new ResponseStatus();
+
+        Connection connection = new OnlineDbConnection().getConnection();
+        JsonNode groupInfo = requestBody.get("object");
+
+        Iterator<Map.Entry<String, JsonNode>> iterator = groupInfo.fields();
+
+        int member_id = Integer.parseInt(iterator.next().toString().split("=")[1]);
+        int groupID = Integer.parseInt(iterator.next().toString().split("=")[1]);
+
+        String leaveGroupQuery = "DELETE FROM `Groupmembers` WHERE `Groupmembers`.`memberID` = ? AND `Groupmembers`.`groupID` = ?";
+        PreparedStatement st = connection.prepareStatement(leaveGroupQuery);
+        st.setInt(1, member_id);
+        st.setInt(2, groupID);
+
+        int count  = st.executeUpdate();
+
+        if(count>0){
+            responseStatus.setStatus(200);
+            responseStatus.setActionToDo("LEAVE GROUP");
+            responseStatus.setMessage("You Left Group Successfully.");
+        }else{
+            responseStatus.setStatus(400);
+            responseStatus.setActionToDo("LEAVE GROUP");
+            responseStatus.setMessage("Unable to LEAVE  group");
+        }
+
+        return new ObjectMapper().writeValueAsString(responseStatus);
+    }
+
+
     public String deleteGroup(JsonNode requestBody) throws Exception{
 
         ResponseStatus responseStatus = new ResponseStatus();
@@ -98,9 +135,10 @@ public class GroupActions {
 
         int group_id = Integer.parseInt(iterator.next().toString().split("=")[1]);
 
-        String deleteGroupQuery = "DELETE FROM `Groups` WHERE `Groups`.`id` = ?";
+        String deleteGroupQuery = "DELETE FROM `Groups` WHERE `Groups`.`id` = ? AND `Groups`.`groupCreatorID` = ?";
         PreparedStatement st = connection.prepareStatement(deleteGroupQuery);
         st.setInt(1, group_id);
+        st.setInt(2, 0);
 
         int count  = st.executeUpdate();
